relocated the either tests to match namespace
